requirements.txt
pandas==2.2.3
pdfplumber==0.11.2
python-dateutil==2.9.0.post0
pytest==8.3.2
typing-extensions==4.12.2

